### Description:

When deleting manufacturing orders, a lot of time is taken by badly optimized code. The method `_action_cancel` and the compute `_compute_picking_ids` are both culprits of this slowness and can be both optimized.

Also, prior to this fix, the test `test_component_and_byproduct_on_transfers` was working by hoping the order of the picking was in a descendant order. However, by batching the `_compute_picking_ids`, the recordset is now automatically in an ascendant order. This PR also change the test to mirror this new behavior.

### Fix:

By extracting the loop used to link `documents` to an MO in `_action_cancel`, we can drastically reduce the time taken by it. However, the main bottleneck was `_compute_picking_ids`, caused by having a search inside a loop. By moving it outside and grouping it by `group_id`, we can avoid doing multiple calls to the ORM.

### Benchmark:

| N° of MO | Before | After |
|----------|--------|-------|
|       80 |    15s |    1s |
|      160 |    33s |    2s |
|     1000 |   2:52 |    8s |

### Reference:

opw-4734710

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr