@robodoo 2025-04-29
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/207960.png)](https://mergebot.odoo.com/odoo/odoo/pull/207960)