@ticodoo 2025-06-16
@robodoo r+

Note that `aggregates=['id:recordset']` was applied so that sorting could be ensured to match original `stock.picking._order` behavior (i.e. so behavior doesn't change in stable)